+++ R/db-compute.R
@@ -70,7 +70,7 @@ db_compute.DBIConnection <- function(con,
     unique_indexes <- as.list(unique_indexes)
   }
 
-  db_save_query(con, sql, table, temporary = temporary)
+  table <- db_save_query(con, sql, table, temporary = temporary)
   db_create_indexes(con, table, unique_indexes, unique = TRUE)
   db_create_indexes(con, table, indexes, unique = FALSE)
   if (analyze) db_analyze(con, table)